Change name for fraudDetection Add header + clean imports
