+++ grc.fraudDetection_web/src/main/java/com/typedb/examples/fraud/model/CreditCard.java
@@ -1,0 +1,81 @@
+/*
+ * Copyright (C) 2022 Vaticle
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.typedb.examples.fraud.model;
+
+import com.opencsv.bean.CsvBindByName;
+import java.util.Objects;
+
+public class CreditCard {
+
+  @CsvBindByName(column = "cc_num")
+  private String number;
+
+  private Bank bank;
+
+  public CreditCard() {}
+
+  public CreditCard(String number, Bank bank) {
+
+    this.number = number;
+    this.bank = bank;
+  }
+
+  public String getNumber() {
+    return number;
+  }
+
+  public Bank getBank() {
+    return bank;
+  }
+
+  public void setBank(Bank bank) {
+    this.bank = bank;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+
+    if (this == o) {
+        return true;
+    }
+    if (!(o instanceof CreditCard that)) {
+        return false;
+    }
+
+    return number.equals(that.number) && bank.equals(that.bank);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(number, bank);
+  }
+
+  @Override
+  public String toString() {
+
+    return
+        "creditCard {" +
+        "  number = '" + number + "', " +
+        "  bank = '" + bank + "'" +
+        "}";
+  }
+}
