+++ grc.fraudDetection_web/src/main/java/com/typedb/examples/fraud/model/MerchantCoordinates.java
@@ -1,0 +1,84 @@
+/*
+ * Copyright (C) 2022 Vaticle
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.typedb.examples.fraud.model;
+
+import com.opencsv.bean.CsvBindByName;
+import java.util.Objects;
+
+public class MerchantCoordinates {
+
+  @CsvBindByName(column = "merch_lat")
+  private String latitude;
+  @CsvBindByName(column = "merch_long")
+  private String longitude;
+
+  public MerchantCoordinates() {}
+
+  public MerchantCoordinates(String latitude, String longitude) {
+
+    this.latitude = latitude;
+    this.longitude = longitude;
+  }
+
+  public String getLatitude() {
+    return latitude;
+  }
+
+  public void setLatitude(String latitude) {
+    this.latitude = latitude;
+  }
+
+  public String getLongitude() {
+    return longitude;
+  }
+
+  public void setLongitude(String longitude) {
+    this.longitude = longitude;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+
+    if (this == o) {
+      return true;
+    }
+    if (!(o instanceof MerchantCoordinates that)) {
+      return false;
+    }
+
+    return latitude.equals(that.latitude) && longitude.equals(that.longitude);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(latitude, longitude);
+  }
+
+  @Override
+  public String toString() {
+    return
+        "merchantCoordinates {" +
+        "  latitude = '" + latitude + "', " +
+        ", longitude = '" + longitude + "'" +
+        "}";
+  }
+}
