+++ grc.fraudDetection_web/src/main/resources/META-INF/resources/index.html
