--- grc.fraud_web/src/main/java/com/typedb/examples/fraud/db/CardholderCoordsDao.java
@@ -1,17 +1,0 @@
-package com.typedb.examples.fraud.db;
-
-import com.typedb.examples.fraud.model.CardholderCoordinates;
-import java.util.Hashtable;
-
-public class CardholderCoordsDao {
-
-  protected static CardholderCoordinates fromResult(Hashtable<String, String> result) {
-
-    var lat = result.get("cardholderLat");
-    var lon = result.get("cardholderLon");
-
-    var coords = new CardholderCoordinates(lat, lon);
-
-    return coords;
-  }
-}
