--- grc.fraud_web/src/main/java/com/typedb/examples/fraud/db/CreditCardDao.java
@@ -1,18 +1,0 @@
-package com.typedb.examples.fraud.db;
-
-import com.typedb.examples.fraud.model.CreditCard;
-import java.util.Hashtable;
-
-public class CreditCardDao {
-
-  protected static CreditCard fromResult(Hashtable<String, String> result) {
-
-    var bank = BankDao.fromResult(result);
-
-    var ccNum = result.get("ccNum");
-
-    var cc = new CreditCard(ccNum, bank);
-
-    return cc;
-  }
-}
