--- grc.fraud_web/src/main/java/com/typedb/examples/fraud/db/MerchantCoordsDao.java
@@ -1,17 +1,0 @@
-package com.typedb.examples.fraud.db;
-
-import com.typedb.examples.fraud.model.MerchantCoordinates;
-import java.util.Hashtable;
-
-public class MerchantCoordsDao {
-
-  protected static MerchantCoordinates fromResult(Hashtable<String, String> result) {
-
-    var lat = result.get("merchantLat");
-    var lon = result.get("merchantLon");
-
-    var coords = new MerchantCoordinates(lat, lon);
-
-    return coords;
-  }
-}
