--- grc.fraud_web/src/main/java/com/typedb/examples/fraud/db/StandardDao.java
@@ -1,12 +1,0 @@
-package com.typedb.examples.fraud.db;
-
-import java.util.Hashtable;
-import java.util.Set;
-
-public interface StandardDao<T> {
-
-  Set<T> getAll();
-
-  void insertAll(Set<T> data);
-
-}
