--- grc.fraud_web/src/main/java/com/typedb/examples/fraud/util/Formatter.java
@@ -1,13 +1,0 @@
-package com.typedb.examples.fraud.util;
-
-public class Formatter {
-
-  public static String formatDateTime(String dateTime) {
-
-    String[] dateTimeParts = dateTime.split(" ");
-
-    var formattedDateTime = dateTimeParts[0] + "T" + dateTimeParts[1];
-
-    return formattedDateTime;
-  }
-}
